+++ src/main/java/com/BookStoreV1/BookStore/Service/UserService.java
@@ -1,13 +1,15 @@
 package com.BookStoreV1.BookStore.Service;
 
+import com.BookStoreV1.BookStore.Validation.User.RelatedUserRent;
 import com.BookStoreV1.BookStore.Validation.User.UserAlreadExistsException;
-import com.BookStoreV1.BookStore.Dto.MessageDTO;
 import com.BookStoreV1.BookStore.Dto.UserDTO;
 import com.BookStoreV1.BookStore.Model.User;
 import com.BookStoreV1.BookStore.Repository.UserRepository;
 import com.BookStoreV1.BookStore.Mapper.UserMapper;
 import com.BookStoreV1.BookStore.Validation.User.UserNotFoundException;
+import lombok.SneakyThrows;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -21,21 +23,24 @@ public class UserService {
 
     @Autowired
     private UserRepository userRepository;
+    @Autowired
+    @Lazy
+    private RentService rentService;
 
-    public MessageDTO create(UserDTO userToCreateDTo){
-        VerifyExistss(userToCreateDTo.getEmail());
-        User userToCreate = userMapper.toModel(userToCreateDTo);
-       User createdUser = userRepository.save(userToCreate);
-       return creationMessage(createdUser);
+    public UserDTO create(UserDTO userDTO){
+        VerifyExistss(userDTO.getEmail());
+        User userToCreate = userMapper.toModel(userDTO);
+        User createdUser = userRepository.save(userToCreate);
+        return userMapper.toDto(createdUser);
     }
 
-    public MessageDTO update(Long id, UserDTO userToUpdateDTo){
+    public UserDTO update(Long id, UserDTO userDTO){
         User foundUser = verifyAndGetUser(id);
 
-        userToUpdateDTo.setId(foundUser.getId());
-        User userToUpdate = userMapper.toModel(userToUpdateDTo);
+        userDTO.setId(foundUser.getId());
+        User userToUpdate = userMapper.toModel(userDTO);
         User updateUser = userRepository.save(userToUpdate);
-        return updateMessage(updateUser);
+        return userMapper.toDto(updateUser);
     }
 
     public UserDTO findById(Long id){
@@ -50,8 +55,12 @@ public class UserService {
                 .collect(Collectors.toList());
     }
 
+    @SneakyThrows
     public void delete(Long id){
         verifyAndGetUser(id);
+        if (rentService.hasRentsForUser(id)) {
+            throw new RelatedUserRent(id);
+        }
         userRepository.deleteById(id);
     }
 
@@ -62,30 +71,11 @@ public class UserService {
         }
     }
 
-    private MessageDTO creationMessage(User createdUser) {
-        return returnMessage(createdUser, "Createdd");
-    }
-    private MessageDTO updateMessage(User updateUser) {
-        return returnMessage(updateUser, "Updatedd");
-    }
-
-    private MessageDTO returnMessage(User updateUser, String action) {
-        String username = updateUser.getNome();
-        Long createdId = updateUser.getId();
-        String createdUserMessage = String.format("UserName %s with ID %s successfuly %s", username, createdId, action);
-        return MessageDTO.builder()
-                .message(createdUserMessage)
-                .build();
-    }
 
     public User verifyAndGetUser(Long id){
         User foundUser = userRepository.findById(id)
                 .orElseThrow(() -> new UserNotFoundException(id));
         return foundUser;
 
-    }
-    private void verifyIfExists(String userEmail){
-        userRepository.findByEmail(userEmail)
-                .ifPresent(user -> {throw new UserAlreadExistsException(userEmail);});
     }
 }
